# Conflicts:
#	src/main/java/team/lodestar/lodestone/modules/rendering/handlers/ParticleHandler.java
#	src/main/java/team/lodestar/lodestone/modules/rendering/particle/pool/ParticlePool.java

+++ src/main/java/team/lodestar/lodestone/helpers/ColorHelper.java
@@ -30,56 +30,56 @@ public class ColorHelper {
         return FastColor.ARGB32.color(a, r, g, b);
     }
 
-    public static int getColor(float r, float g, float b, float a) {
-        return FastColor.ARGB32.color((int) (a * 255f), (int) (r * 255f), (int) (g * 255f), (int) (b * 255f));
+    public static int getColor(double r, double g, double b, double a) {
+        return FastColor.ARGB32.color((int) (a * 255), (int) (r * 255), (int) (g * 255), (int) (b * 255));
     }
 
-    public static Color colorLerp(Easing easing, float delta, Color startColor, Color endColor) {
+    public static Color colorLerp(Easing easing, double delta, Color startColor, Color endColor) {
         delta = Mth.clamp(delta, 0, 1);
         int br = startColor.getRed(), bg = startColor.getGreen(), bb = startColor.getBlue();
         int dr = endColor.getRed(), dg = endColor.getGreen(), db = endColor.getBlue();
-        float eased = easing.ease(delta);
+        double eased = easing.ease(delta);
         int red = (int) Mth.lerp(eased, br, dr);
         int green = (int) Mth.lerp(eased, bg, dg);
         int blue = (int) Mth.lerp(eased, bb, db);
         return new Color(Mth.clamp(red, 0, 255), Mth.clamp(green, 0, 255), Mth.clamp(blue, 0, 255));
     }
 
-    public static Color colorLerp(Easing easing, float delta, float min, float max, Color startColor, Color endColor) {
+    public static Color colorLerp(Easing easing, double delta, double min, double max, Color startColor, Color endColor) {
         delta = Mth.clamp(delta, 0, 1);
         int br = startColor.getRed(), bg = startColor.getGreen(), bb = startColor.getBlue();
         int dr = endColor.getRed(), dg = endColor.getGreen(), db = endColor.getBlue();
-        float eased = easing.lerp(delta, min, max);
+        double eased = easing.lerp(delta, min, max);
         int red = (int) Mth.lerp(eased, br, dr);
         int green = (int) Mth.lerp(eased, bg, dg);
         int blue = (int) Mth.lerp(eased, bb, db);
         return new Color(Mth.clamp(red, 0, 255), Mth.clamp(green, 0, 255), Mth.clamp(blue, 0, 255));
     }
 
-    public static Color multicolorLerp(Easing easing, float delta, Color... colors) {
+    public static Color multicolorLerp(Easing easing, double delta, Color... colors) {
         return multicolorLerp(easing, delta, List.of(colors));
     }
 
-    public static Color multicolorLerp(Easing easing, float delta, List<Color> colors) {
+    public static Color multicolorLerp(Easing easing, double delta, List<Color> colors) {
         delta = Mth.clamp(delta, 0, 1);
         int colorCount = colors.size() - 1;
-        float eased = easing.ease(delta);
-        float colorIndex = colorCount * eased;
+        double eased = easing.ease(delta);
+        double colorIndex = colorCount * eased;
         int index = (int) Mth.clamp(colorIndex, 0, colorCount);
         Color color = colors.get(index);
         Color nextColor = index == colorCount ? color : colors.get(index + 1);
         return ColorHelper.colorLerp(easing, colorIndex - (int) (colorIndex), color, nextColor);
     }
 
-    public static Color multicolorLerp(Easing easing, float pct, float min, float max, Color... colors) {
+    public static Color multicolorLerp(Easing easing, double pct, double min, double max, Color... colors) {
         return multicolorLerp(easing, pct, min, max, List.of(colors));
     }
 
-    public static Color multicolorLerp(Easing easing, float delta, float min, float max, List<Color> colors) {
+    public static Color multicolorLerp(Easing easing, double delta, double min, double max, List<Color> colors) {
         delta = Mth.clamp(delta, 0, 1);
         int colorCount = colors.size() - 1;
-        float eased = easing.ease(delta);
-        float colorIndex = colorCount * eased;
+        double eased = easing.ease(delta);
+        double colorIndex = colorCount * eased;
         int index = (int) Mth.clamp(colorIndex, 0, colorCount);
         Color color = colors.get(index);
         Color nextColor = index == colorCount ? color : colors.get(index + 1);
@@ -90,8 +90,8 @@ public class ColorHelper {
         return darker(color, times, 0.7f);
     }
 
-    public static Color darker(Color color, int power, float factor) {
-        float FACTOR = (float) Math.pow(factor, power);
+    public static Color darker(Color color, int power, double factor) {
+        double FACTOR = Math.pow(factor, power);
         return new Color(Math.max((int) (color.getRed() * FACTOR), 0),
                 Math.max((int) (color.getGreen() * FACTOR), 0),
                 Math.max((int) (color.getBlue() * FACTOR), 0),
@@ -102,8 +102,8 @@ public class ColorHelper {
         return brighter(color, power, 0.7f);
     }
 
-    public static Color brighter(Color color, int power, float factor) {
-        float FACTOR = (float) Math.pow(factor, power);
+    public static Color brighter(Color color, int power, double factor) {
+        double FACTOR = Math.pow(factor, power);
         int r = color.getRed();
         int g = color.getGreen();
         int b = color.getBlue();
